Most of the demand forcasting module is completed
It basically forcast the future demand and help suppiler to adjust the stock according to the market need. Most of the part is completed of this module. There still other module remain which was depend on this module so that why I have to take this module in work first still it is under progress in future you will get more refine module.So the next module we have to take Inventary Management which depend on demand forcast module After completing all module we can make a micro service and make it independent module and loose couple for future work